Fix order in which installation prefix is set
The installation prefix is overwritten in case the user does not
specify and installation prefix themselves. This commit makes sure
that the prefix is overwritten before the GNUInstallDirs macro
is included. This avoids problems with some CMake variable containing
the default values of the GNUInstallDirs although the CMAKE_PREFIX_PATH
has been overwritten. This was a problem, e.g., for the pkg-config file
generation.