Fix compilation with disabled Curl/Azure Presigned
When DISABLE_AZURE_PRESIGNED_IOMANAGER / DISABLE_CURL_IOMANAGER are set
to on, Debug build fails while Release builds fine.

This seems to be caused by LTO being enabled by default for Release
builds and disabled for Debug ones.

The actual issue is that additional source files have to be
marked as "header file only" for linking to succeed.